Made Major Improvements To The Backward Pass
Many of the functions within the backward pass were wrong which caused the model to not learn at all. With the corrections I have made, the model is able to somewhat learn. This change can be seen as the loss usually decreases at each step in the learning process. Other functions within the code were also changed to help improve the backward pass. Additionally, the weights and biases are randomly sampled from a normal distribution rather than a uniform distribution.